+++ Sources/KeyboardKit/System/SystemKeyboardButton.swift
@@ -34,6 +34,6 @@ public struct SystemKeyboardButton: View {
     public var body: some View {
         SystemKeyboardButtonContent(action: action, text: text, image: image)
             .standardButtonStyle(for: action, context: context)
-            .keyboardAction(action, actionHandler: context.actionHandler)
+            .keyboardGestures(for: action, actionHandler: context.actionHandler)
     }
 }